Rename measure and start improving tests.
