--- lib/measures/GEBAppliancesPeakPeriodShift/README.md
@@ -1,52 +1,0 @@
-
-
-###### (Automatically generated documentation)
-
-# GEBAppliancesPeakPeriodShift
-
-## Description
-TODO
-
-## Modeler Description
-TODO
-
-## Measure Type
-ModelMeasure
-
-## Taxonomy
-
-
-## Arguments
-
-
-### Schedules: Peak Period
-Specifies the peak period. Enter a time like "15 - 18" (start hour can be 0 through 23 and end hour can be 1 through 24).
-**Name:** schedules_peak_period,
-**Type:** String,
-**Units:** ,
-**Required:** true,
-**Model Dependent:** false
-
-
-### Schedules: Peak Period Delay
-The number of hours after peak period end.
-**Name:** schedules_peak_period_delay,
-**Type:** Integer,
-**Units:** hr,
-**Required:** true,
-**Model Dependent:** false
-
-
-### Schedules: Peak Period Schedule File Column Names
-Comma-separated list of schedule file column names to shift during the peak period.
-**Name:** schedules_peak_period_column_names,
-**Type:** String,
-**Units:** ,
-**Required:** false,
-**Model Dependent:** false
-
-
-
-
-
-
