+++ lib/measures/PeakPeriodSchedulesShift/measure.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0"?>
 <measure>
   <schema_version>3.0</schema_version>
-  <name>geb_appliances_peak_period_shift</name>
+  <name>peak_period_schedules_shift</name>
   <uid>4954275f-9caf-422c-9e1f-061244ade152</uid>
-  <version_id>9a35a61d-f71d-431b-af46-8e431be663af</version_id>
-  <version_modified>20230330T215718Z</version_modified>
+  <version_id>4d978b1b-4c0f-4911-9cbd-7db53e8d167a</version_id>
+  <version_modified>20230331T220130Z</version_modified>
   <xml_checksum>A8988589</xml_checksum>
-  <class_name>GEBAppliancesPeakPeriodShift</class_name>
-  <display_name>GEBAppliancesPeakPeriodShift</display_name>
-  <description>TODO</description>
-  <modeler_description>TODO</modeler_description>
+  <class_name>PeakPeriodSchedulesShift</class_name>
+  <display_name>PeakPeriodSchedulesShift</display_name>
+  <description>Shifts select schedules out of a peak period.</description>
+  <modeler_description>Enter a weekday peak period, a delay value, and any applicable ScheduleRuleset or ScheduleFile schedules. Shift all schedule values falling within the peak period to after the end (offset by delay) of the peak period. Prevent stacking of schedule values by only allowing shifts to all-zero periods.</modeler_description>
   <arguments>
     <argument>
       <name>schedules_peak_period</name>
@@ -31,9 +31,17 @@
       <default_value>0</default_value>
     </argument>
     <argument>
-      <name>schedules_peak_period_column_names</name>
-      <display_name>Schedules: Peak Period Schedule File Column Names</display_name>
-      <description>Comma-separated list of schedule file column names to shift during the peak period.</description>
+      <name>schedules_peak_period_schedule_rulesets_names</name>
+      <display_name>Schedules: Peak Period Schedule Rulesets Names</display_name>
+      <description>Comma-separated list of Schedule:Ruleset object names corresponding to schedules to shift during the specified peak period.</description>
+      <type>String</type>
+      <required>false</required>
+      <model_dependent>false</model_dependent>
+    </argument>
+    <argument>
+      <name>schedules_peak_period_schedule_files_column_names</name>
+      <display_name>Schedules: Peak Period Schedule Files Column Names</display_name>
+      <description>Comma-separated list of column names, referenced by Schedule:File objects, corresponding to schedules to shift during the specified peak period.</description>
       <type>String</type>
       <required>false</required>
       <model_dependent>false</model_dependent>
@@ -70,16 +78,34 @@
       <checksum>CD7F5672</checksum>
     </file>
     <file>
+      <filename>files/schedules20230327-33728-y5y5a5.csv</filename>
+      <filetype>csv</filetype>
+      <usage_type>test</usage_type>
+      <checksum>DD03C7FA</checksum>
+    </file>
+    <file>
+      <filename>README.md</filename>
+      <filetype>md</filetype>
+      <usage_type>readme</usage_type>
+      <checksum>F02F0326</checksum>
+    </file>
+    <file>
       <filename>base-schedules-detailed-occupancy-stochastic.osm</filename>
       <filetype>osm</filetype>
       <usage_type>test</usage_type>
-      <checksum>AC4D8730</checksum>
+      <checksum>ED297470</checksum>
     </file>
     <file>
-      <filename>test_GEBAppliancesPeakPeriodShift.osm</filename>
+      <filename>PeakPeriodSchedulesShift_Test.rb</filename>
+      <filetype>rb</filetype>
+      <usage_type>test</usage_type>
+      <checksum>F412B400</checksum>
+    </file>
+    <file>
+      <filename>test_PeakPeriodSchedulesShift.osm</filename>
       <filetype>osm</filetype>
       <usage_type>test</usage_type>
-      <checksum>AC4D8730</checksum>
+      <checksum>B31A5FC9</checksum>
     </file>
     <file>
       <version>
@@ -90,25 +116,7 @@
       <filename>measure.rb</filename>
       <filetype>rb</filetype>
       <usage_type>script</usage_type>
-      <checksum>1B30550C</checksum>
-    </file>
-    <file>
-      <filename>README.md</filename>
-      <filetype>md</filetype>
-      <usage_type>readme</usage_type>
-      <checksum>F5D76499</checksum>
-    </file>
-    <file>
-      <filename>files/schedules20230327-33728-y5y5a5.csv</filename>
-      <filetype>csv</filetype>
-      <usage_type>test</usage_type>
-      <checksum>748724F1</checksum>
-    </file>
-    <file>
-      <filename>GEBAppliancesPeakPeriodShift_Test.rb</filename>
-      <filetype>rb</filetype>
-      <usage_type>test</usage_type>
-      <checksum>81A4B74E</checksum>
+      <checksum>4B851402</checksum>
     </file>
   </files>
 </measure>
