+++ lib/measures/PeakPeriodSchedulesShift/tests/base-schedules-detailed-occupancy-stochastic.osm
@@ -3607,19 +3607,19 @@ OS:Schedule:Day,
   0.763686131386861,                      !- Value Until Time 14
   20,                                     !- Hour 15
   0,                                      !- Minute 15
-  0.733138686131387,                      !- Value Until Time 15
+  0,                                      !- Value Until Time 15
   21,                                     !- Hour 16
   0,                                      !- Minute 16
-  0.717864963503649,                      !- Value Until Time 16
+  0,                                      !- Value Until Time 16
   22,                                     !- Hour 17
   0,                                      !- Minute 17
-  0.702591240875912,                      !- Value Until Time 17
+  0,                                      !- Value Until Time 17
   23,                                     !- Hour 18
   0,                                      !- Minute 18
-  0.672043795620438,                      !- Value Until Time 18
+  0,                                      !- Value Until Time 18
   24,                                     !- Hour 19
   0,                                      !- Minute 19
-  0.626222627737226;                      !- Value Until Time 19
+  0;                                      !- Value Until Time 19
 
 OS:Schedule:Rule,
   {7d5c0aeb-fc0d-4f89-9915-c4e207854d15}, !- Handle
